api/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
psycopg2-binary==2.9.9
redis==5.0.3
passlib==1.7.4
bcrypt==3.2.2